+++ src/main/kotlin/io/nofrills/multimodule/BasePlugin.kt
@@ -16,7 +16,6 @@
 
 package io.nofrills.multimodule
 
-import org.gradle.api.Action
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.artifacts.ExternalModuleDependency
@@ -27,7 +26,6 @@ import org.gradle.api.publish.PublishingExtension
 import org.gradle.api.publish.maven.MavenPublication
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.api.tasks.bundling.Jar
-import org.gradle.testing.jacoco.tasks.JacocoReport
 import org.jetbrains.kotlin.gradle.plugin.getKotlinPluginVersion
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
@@ -61,7 +59,7 @@ abstract class BasePlugin : Plugin<Project> {
     }
 
     /** Apply the jacoco plugin. */
-    protected abstract fun applyJacoco(project: Project, jacocoAction: Action<JacocoReport>)
+    protected abstract fun applyJacoco(project: Project, jacocoConfig: JacocoConfig)
 
     /** Apply the kotlin plugin appropriate for the module. */
     protected abstract fun applyKotlin(project: Project, kotlinConfig: KotlinConfig)
@@ -125,7 +123,9 @@ abstract class BasePlugin : Plugin<Project> {
         multimoduleExtension.jacocoAction?.let { jacocoAction ->
             project.afterEvaluate { project ->
                 if (submoduleExtension.jacocoAllowed.get()) {
-                    applyJacoco(project, jacocoAction)
+                    val jacocoConfig = JacocoConfig()
+                    jacocoAction.execute(jacocoConfig)
+                    applyJacoco(project, jacocoConfig)
                 }
             }
         }